Trim down Pivot tests to useful cases, rename new implementations, clean up.
