--- graylog-storage-opensearch2/src/test/java/org/graylog/storage/opensearch2/views/searchtypes/pivots/buckets/OSTimeHandlerTest.java
@@ -1,106 +1,0 @@
-/*
- * Copyright (C) 2020 Graylog, Inc.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the Server Side Public License, version 1,
- * as published by MongoDB, Inc.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * Server Side Public License for more details.
- *
- * You should have received a copy of the Server Side Public License
- * along with this program. If not, see
- * <http://www.mongodb.com/licensing/server-side-public-license>.
- */
-package org.graylog.storage.opensearch2.views.searchtypes.pivots.buckets;
-
-import org.graylog.plugins.views.search.Query;
-import org.graylog.plugins.views.search.searchtypes.pivot.Pivot;
-import org.graylog.plugins.views.search.searchtypes.pivot.buckets.DateInterval;
-import org.graylog.plugins.views.search.searchtypes.pivot.buckets.Interval;
-import org.graylog.plugins.views.search.searchtypes.pivot.buckets.Time;
-import org.graylog.plugins.views.search.timeranges.DerivedTimeRange;
-import org.graylog.storage.opensearch2.views.OSGeneratedQueryContext;
-import org.graylog.storage.opensearch2.views.searchtypes.pivot.AggTypes;
-import org.graylog.storage.opensearch2.views.searchtypes.pivot.OSPivot;
-import org.graylog.storage.opensearch2.views.searchtypes.pivot.buckets.OSTimeHandler;
-import org.graylog2.plugin.indexer.searches.timeranges.InvalidRangeParametersException;
-import org.graylog2.plugin.indexer.searches.timeranges.RelativeRange;
-import org.graylog2.plugin.indexer.searches.timeranges.TimeRange;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
-import org.mockito.ArgumentCaptor;
-
-import java.util.Optional;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.RETURNS_DEEP_STUBS;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
-
-class OSTimeHandlerTest {
-    private OSTimeHandler esTimeHandler;
-
-    private final Pivot pivot = mock(Pivot.class);
-
-    private final Time time = mock(Time.class);
-
-    private final OSPivot esPivot = mock(OSPivot.class);
-
-    private final OSGeneratedQueryContext queryContext = mock(OSGeneratedQueryContext.class, RETURNS_DEEP_STUBS);
-
-    private final Query query = mock(Query.class);
-
-    private final Interval interval = mock(Interval.class, RETURNS_DEEP_STUBS);
-
-    @BeforeEach
-    public void setUp() throws Exception {
-        this.esTimeHandler = new OSTimeHandler();
-        when(time.interval()).thenReturn(interval);
-        when(time.field()).thenReturn("foobar");
-        final AggTypes aggTypes = mock(AggTypes.class);
-        when(queryContext.contextMap().get(any())).thenReturn(aggTypes);
-        when(query.effectiveTimeRange(any())).thenCallRealMethod();
-    }
-
-    @Test
-    public void timeSpecIntervalIsCalculatedOnPivotTimerangeIfOverridden() throws InvalidRangeParametersException {
-        final ArgumentCaptor<TimeRange> timeRangeCaptor = ArgumentCaptor.forClass(TimeRange.class);
-        when(interval.toDateInterval(timeRangeCaptor.capture())).thenReturn(DateInterval.days(1));
-        when(pivot.timerange()).thenReturn(Optional.of(DerivedTimeRange.of(RelativeRange.create(4242))));
-
-        this.esTimeHandler.doCreateAggregation("foobar", pivot, time, queryContext, query);
-
-        final TimeRange argumentTimeRange = timeRangeCaptor.getValue();
-        assertThat(argumentTimeRange).isEqualTo(RelativeRange.create(4242));
-    }
-
-    @Test
-    public void timeSpecIntervalIsCalculatedOnQueryTimeRangeIfNoPivotTimeRange() throws InvalidRangeParametersException {
-        final ArgumentCaptor<TimeRange> timeRangeCaptor = ArgumentCaptor.forClass(TimeRange.class);
-        when(interval.toDateInterval(timeRangeCaptor.capture())).thenReturn(DateInterval.days(1));
-        when(pivot.timerange()).thenReturn(Optional.empty());
-        when(query.timerange()).thenReturn(RelativeRange.create(2323));
-
-        this.esTimeHandler.doCreateAggregation("foobar", pivot, time, queryContext, query);
-
-        final TimeRange argumentTimeRange = timeRangeCaptor.getValue();
-        assertThat(argumentTimeRange).isEqualTo(RelativeRange.create(2323));
-    }
-
-
-    @ParameterizedTest
-    @ValueSource(strings = { "1s", "1M", "4s", "14d" })
-    public void correctIntervalTypeIsUsedForAggregation(String intervalString) throws InvalidRangeParametersException {
-        when(pivot.timerange()).thenReturn(Optional.empty());
-        when(query.timerange()).thenReturn(RelativeRange.create(2323));
-        when(interval.toDateInterval(any(TimeRange.class)).toString()).thenReturn(intervalString);
-
-        this.esTimeHandler.doCreateAggregation("foobar", pivot, time, queryContext, query);
-    }
-}
